+++ docs/source/whatsnew.rst
@@ -3,6 +3,15 @@
 What's New
 ==========
 
+v4.5.1
+------
+- Devices has been revamped to handle PhoneOS devices as well as prepare the code for upcoming changes to /v1/devices
+  - :class:`~.device.DeviceList` now always uses `/v1/devices` regardless of whether the data is being requested at the Org, Location, Person or Workspace level
+  - :class:`~.device.Device` now has a :attr:`~.device.Device.calling_id` property, which is the Webex Calling ID used for all of the `/v1/telephony` API endpoints
+  - One behavior that changes in this design is that a the :attr:`Person.devices <.person.Person.devices>` now only returns devices that are owned by the user.
+  - A new property :attr:`Person.appearance_devices <.person.Person.appearance_devices>` has been added to find all the devices the user has a line appearance on
+- :class:`~.virtual_line.VirtualLine` now has an :attr:`~.virtual_line.VirtualLine.name` attribute
+
 v4.5.0
 ------
 - The inclusion of the `meraki` module as a dependency doesn't make sense for 99% of **wxcadm** users, so I have moved it to an optional dependency. See :ref:`Using Meraki for RedSky Network Discovery`